+++ library/src/linkbot.cpp
@@ -276,7 +276,7 @@ void Linkbot::getJointSafetyThresholds(int& t1, int& t2, int& t3)
     }
 }
 
-void Linkbot::getJointSafetyAngles(float& t1, float& t2, float& t3)
+void Linkbot::getJointSafetyAngles(double& t1, double& t2, double& t3)
 {
     try {
         auto value = asyncFire(m->robot,
